requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
dbt-core>=1.7.0
dbt-snowflake>=1.7.0
astronomer-cosmos>=1.2.0
snowflake-connector-python[pandas]
apache-airflow-providers-snowflake 